+++ scripts/textura_2.gd
@@ -441,8 +441,9 @@ func handle_hurt_animation() -> void:
 	deactivate_all_attack_areas()
 	deactivate_all_special_attack_areas()
 
-	if sprite_frames.has_animation(&"hurt"):
-		play_if_different(&"hurt")
+	var hurt_anim := &"hurt_super" if player.form == player.Form.SUPER else &"hurt"
+	if sprite_frames.has_animation(hurt_anim):
+		play_if_different(hurt_anim)
 		return
 
 	handle_death_animation()